Move gui_score_editor into widgets/
