+++ include/fullscore/actions/set_normal_mode_action.h
@@ -5,7 +5,7 @@
 
 #include <fullscore/actions/action_base.h>
 
-#include <fullscore/gui_score_editor.h>
+#include <fullscore/widgets/gui_score_editor.h>
