+++ include/fullscore/widgets/gui_score_editor.h
