+++ src/actions/move_cursor_left_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/move_cursor_left_action.h>
 
-#include <fullscore/gui_score_editor.h>
+#include <fullscore/widgets/gui_score_editor.h>
