+++ src/actions/toggle_show_debug_data_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/toggle_show_debug_data_action.h>
 
-#include <fullscore/gui_score_editor.h>
+#include <fullscore/widgets/gui_score_editor.h>
