+++ src/widgets/gui_score_editor.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/gui_score_editor.h>
+#include <fullscore/widgets/gui_score_editor.h>
 
 #include <allegro_flare/allegro_flare.h>
